requirements.txt
amqp==5.0.6
argon2-cffi==20.1.0
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
billiard==3.6.4.0
bleach==4.0.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.6
django-autoslug==1.9.8
django-cors-headers==3.7.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
dnspython==2.1.0
drf-yasg==1.20.0
entrypoints==0.3
eventlet==0.32.0
flake8==3.9.2
flower==1.0.0
greenlet==1.1.1
humanize==3.11.0
idna==3.2
inflection==0.5.1
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.1
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kombu==5.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
oauthlib==3.1.1
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytils==0.3
pytz==2021.1
pywin32==301
pywinpty==1.1.3
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.10.0
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
send-email==201212
Send2Trash==1.8.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
stripe==2.60.0
terminado==0.11.1
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
